Add ownercompanie field to User and update DataEditor
Introduces the ownercompanie field to the User type and updates DataEditor to handle and display this field. The UI now shows the owning company name if present and allows selection of the company by name. The location field has been removed from the user form as part of this update.